requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
opencv-python-headless==4.5.5.62
ultralytics==8.0.0
pillow==9.0.0
jinja2==3.1.2
python-multipart==0.0.6  # Add this line
